Match indent of the previous line…
…when there's no indents query to help us out.

This seems to be the default behavior for plaintext and for other language
modes.

I've got a Tree-sitter Markdown grammar locally and I noticed it wasn't
preserving indentation when I was two levels deep into a bulleted list and hit
`Enter` — I had to re-indent each time. (The Markdown grammar doesn't include an
indents query because there's practically no way to predict indentation levels
in Markdown.)