Change `injectionLayersAtPoint`…
…to consider content ranges instead of extents.

If you want to know which layers are operative at a given point, it makes no
sense to check a layer's extent. A `LanguageLayer` could extend over the entire
buffer and still only apply highlighting to one very small range in the buffer.

Thus it makes no sense for `controllingLayerAtPoint` not to check a layer's
content ranges, nor does it make sense for the `injectionLayersAtPoint` method
that it relies on.

If I felt like there was any use in knowing that a language layer extended over
a range that included the given point but wasn't operative at that point, I'd be
hesitant to change these methods. But I don't.

This bug was discovered when I realized that the wrong indents query was being
consulted in the following HTML…

    <style>
      #foo {

…when I moved the cursor to the end of the `<style>` element and hit `Enter`.