Roll back `languageLayersAtPoint` change…
…as overly broad.

Turns out there are some scenarios where you'd want to know about a layer whose
extent may include a given point, even if its content ranges _don't_ include
that point. I'd have realized this if I'd done something as simple as run the
test suite. Silly me.

But `controllingLayerAtPoint` will still enforce the content range constraint
because we use it to pick a winner for indent queries, fold queries, and so on.