Ensure `LanguageLayer#update` performs at least…
…one update.

This commit changes a `while` loop to a `do`/`while` loop so that it’s
guaranteed to run at least once. It’s possible for this loop to repeat an
arbitrary number of times to account for changes made by the user in the middle
of a previous `update` cycle, but it needs to run at least once. Hence the
original loop was improperly structured — if we get to the `update` method, we
need an update to happen.

The scenario that revealed this bug was as follows: a change made on a root
`LanguageLayer` changed the node set of an existing injection layer, but none
of the conditions of the `while` loop were ever true, so the loop was skipped
altogether.

I cannot think of a good test case for this, but all existing tests pass.