Downgrade playwright.
Why? Because it is not running on the CI. But it runs local. Why? I have
no clue. I am _so tired_ of these things that JUST FAIL on the CI...
anyway, I hope I don't commit this and it fails to load Pulsar in the
CI, because there's a possibility this will happen, and I will have no
idea why, because, again, these stupid Javascript tools decide to
swallow logs and stacktraces and exceptions, meaning that when something
fails, you have to rely on crystal balls, magic, and maybe computer
telepathy to discover why.