Transform to open directory as last change
The way this if else is made, is structured quite weirdly. But leaving it as is, to avoid complexity. We now as only a last resort if no other changes are applied, will swap to the open directory icon, this ensures we don't swap a book or repo icon with the open directory