[tree-sitter] Update tree-sitter-typescript
TypeScript keeps adding new syntax, so I've built new WASM files from master
for both `tree-sitter-typescript` and `tree-sitter-tsx`.

The one change that I noticed (and the impetus for updating) is proper
highlighting of the `satisfies` keyword that was added in TS 4.9.

I'm getting tired of making near-identical changes to the `highlights.scm` files
for the two different TypeScript parsers, so I should really figure out how to
make it so that two different grammars can include the same SCM file while still
making it possible to customize the capture names.