[tree-sitter] Add query file concatenation
As part of overhauling the TypeScript grammars — which involve two nearly
identical parsers, one that supports JSX and one that does not — I've made it
possible to specify more than one SCM file for a `highlightsQuery` value (or for
any other kind of query). The files will be concatenated in order, and a newline
will be inserted between each one. This lets the two parsers share most of their
query logic and eliminates a lot of duplication.

In addition, a new magical scope segment has been introduced called _LANG_. It
allows two different grammars to rely upon the same base SCM file while
customizing the scopes it uses. Hence the TypeScript-with-JSX grammar can end
its scope names with `.ts.tsx`, and the TypeScript-by-itself grammar can end
its scope names with `.ts`, even as they both use the same base
`highlights.scm`.

The application for this functionality is pretty limited, because it presumes
a situation where two Tree-sitter parsers will share a core set of parsing
logic, node names, and other conventions. But all that might be applicable for
a parser that extends HTML or CSS parsing.

It's tempting to add a feature that would allow a grammar to build upon the SCM
file from another package's grammar, but it could also be a total disaster,
since it could break whenever that other package upgraded its parser.