Add the Tree-sitter Markdown grammar.
Quite proud of this one — it uses an outer parser to identify front-matter and
highlight it properly, then injects the Markdown parser into everything that
isn't front matter.

The scope names are quite different — because the scope names in the legacy
Markdown grammar are awful — so I adapted the eight built-in syntax themes so
that there are no major style differences.