script: Update version string check for binary renaming script
Need to check this way, since the CI script already modifies
the version string to end in something like "2023081600", not "-dev",
for Rolling releases (or for all builds other than Regular releases).

Also: Update and add more logging of what the script is doing.

Also also: Fix some spaces --> tabs on a couple of lines.