Make the `fun` snippet work in more contexts
`fun` expands to a simple function definition in JavaScript, but there 
are certain contexts where that syntax is invalid, making the snippet 
much less useful.

Now that we're scoping the insides of class bodies and object literals, 
though, we can define more specific snippets for those contexts that use 
the correct syntax.

The class-instance-method snippet and the inside-an-object-literal 
snippet are identical, except that the latter has a trailing comma.