CI: Set CPU limits for Cirrus, set greedy to true
See if we can reduce the CPU-hours, especially billable ones,
that out builds use on Cirrus.

We can potentially get spare/idle resources already privisioned for
Cirrus by setting the "greedy: true" option.

See this blog post for more details:
https://medium.com/cirruslabs/introducing-greedy-container-instances-29aad06dc2b4

So, with the "greedy" config set to "true", we may actually be able to
get back more CPU hours, some of the time, without costing Cirrus as
much, and thereby allowing Cirrus to bill _us_ less.

Note that we are still on the free pseudo-unlimited tier at the
moment, and exploring this option doesn't mean we've ever been billed
by Cirrus (we haven't), nor that we've finalized any decisions about
which CI provider(s) we will/won't be using in the future (still being
discussed and decided ongoing).