script: Use process.exit() in renaming script
Gets rid of a lot of if-else that made the code harder to follow.

This should be pretty safe and unproblematic in a standalone script
like this, versus some more complex or nested Node process scenario,
or scenarios with async input/output that might be ongoing
when the process.exit() is handled.

Also, update some of the code comments a bit.

Co-authored-by: Andrew Dupont <github@andrewdupont.net>