CI: Add binary renaming script to GitHub Actions
This script renames the Pulsar binaries to use the naming scheme
we use for all of the Regular release binaries.

It was already added to the Cirrus config,
now it's used in the GitHub Actions workflow as well.

(It checks the version string in package.json to know not to rename
the binaries unless we're actually trying to make a Regular release.)