settings-view: Don't fix repository for core themes
Stop inserting a best-guess `repository` field for built-in themes'
package.json data if they lack one in their actual package.json file.

Follow-up to https://github.com/pulsar-edit/pulsar/pull/264.

Doing this for the themes panel as well, since we just did it for the
installed-packages panel, and since it's good to be consistent.