CI: Check for forks in a more semantic way for macOS signing
Check "if pull request came from a fork", rather than checking "if the
repo the pull request is from is 'pulsar-edit/pulsar'".

This makes it more friendly to anyone who wishes to fork
pulsar-edit/pulsar and expects to run CI
and have pull requests at said fork.

This is also somewhat shorter and hopefully slightly easier to read.