CI: Set a custom clone script (exit with status 1)
Hopefully, this fails the intentionally-mark-as-failed task
just a little bit faster, saving some billable seconds.

At minimum this should skip any time that would have been wasted on
trying to clone the repo when we don't actually need or want to clone.
(Since we're just trying to fail a task (and the build) as soon as
physically possible, in this case.)