CI: Use commit date as the Rolling version number
The Rolling version number was previously set as the result of a
`date` command, implicitly using server time (as of the time the CI
build was running, in the time zone of the server), meaning the
version number could be different based on server timezone across many
regions the CI could happen to run in, based on the timing of re-runs,
or change due to slight differences in how runners were allocated to
run each job.

Most notably, this could lead to the different OS/arch builds having
different version numbers from the same overall CI build, since they
ran in separate tasks!

Now that we have multiple CI providers building the editor binaries,
we have rather drastically different timing of these runs, and the
version numbers are likely never going to sync up.

---

By setting the version number from the commit date, always in UTC,
Rolling versions on Cirrus should now always sync up
with an existing version as published from GitHub Actions.

This also means we can check the commit dates of recent commits to
`master` branch in order to know which exact commit a given Rolling
release version corresponds to.

---

This idea was worked on by a few people, and also thanks to this
StackOverflow answer:
https://stackoverflow.com/questions/21363187/git-show-dates-in-utc

Co-authored-by: Meadowsys <blazeykirin@gmail.com>
Co-authored-by: confused-Techie <dev@lhbasics.com>