CI: Update code signing env var in GitHub Actions
This is for newer electron-builder. This needs to be set, or else the
latest electron-builder won't attempt to sign/notarize for
Pull Requests.

We want to at least try this once, to try signing in this PR, and see
if the latest electron-builder version fixes the signing/notarizing
issue we have been seeing lately from our binaries built on
GitHub Actions?