CI: Install zstd in Docker so that caches will match
Caches have a unique "version" that relates to the OS/compression
strategy.

zstd is the default for GitHub-hosted Linux runners.

For "self-hosted" or custom runners (or Docker containers, I guess?),
you have to ensure you have zstd on the system creating the cache,
or it won't be restorable in jobs running on GitHub-hosted Linux runners.

So.. Add zstd to the Docker Debian environment so we get compatible caches.