[tree-sitter] Fix faulty assumption about ordering of `findMarkers` results
When determining the possible breadth of injection layers that could be affected by a buffer change, I assumed that `MarkerLayer#findMarkers` would return results such that I could take the first result's start and the last result's end and end up with a range that containd all the results.

#756 proved that not to be true, at least in the case of Clojure — the effect of which is that lots of injection layers were getting ignored during the step where we decide whether to reuse existing layers or create new ones. That led to an ever-increasing number of injection layers being created.

Instead, we need to find the minimum and maximum points in the set and use that as the range.