[tree-sitter] Ensure that candidate injection range doesn’t shrink
When a region of the buffer changes, we want to consider all injections that touch that range, even if they aren't fully enclosed by the range. So we need to be able to grow the original range when searching for injection layers.

This fixes a bug in which the candidate range could inadvertently _shrink_.