fix(tree-sitter): pass node text to grammar
This fixes a subtle breaking change in the WASM tree-sitter modes that
affected the semanticolor package. (And probably only that package.)

Although the node text is unused in `idForScope` in the core editor, the
legacy tree-sitter mode was specifically modified to pass the node text
into `idForScope` so that the semanticolor package could make use of it
to give custom scopes (and thus colors) to different symbols. (On
activation, semanticolor monkey patches that method in relevant
grammars.)

This alone is not enough to get sematicolor working with Pulsar's new
WASM tree-sitter modes, but it's one piece of the puzzle.

Ref: 97b905a2b054c372252a6646e5649eb487487b36
Ref: https://github.com/atom/atom/pull/20212
Ref: https://github.com/pulsar-edit/pulsar/blob/1bf7fc4025f0b10623058e4e8340e166954ed188/src/tree-sitter-language-mode.js#L1321-L1324
Ref: https://github.com/digitallyserviced/semanticolor/blob/d219ef9a1f8fa5251025b74dc6dc0a7f7c38d102/lib/semanticolor-grammar.js#L44-L65