Update the Crowdin URL in README to direct to the project page
This commit updates the Crowdin URL in the README.md file to point
directly to the Pulsar project's page on Crowdin. Previously, the link
redirected users to the Crowdin homepage((https://crowdin.com/)), which
could cause confusion and an extra step for contributors looking to help
with translations. The new URL provides a more streamlined and
user-friendly experience for contributors.