Debugging when a package service is incorrect
This PR is basically to add some debugging info for providers and
consumers. Currently, when a package registers as a service provider or
a service consumer, and doesn't actually implements the right interface
(a function, basically) the package loader silently ignores it. This
commit adds a warning to console so that we know a package did the wrong
thing.