electron-builder: Add '--no-sandbox' launch arg for Linux build targets
Fixes Dev Tools not being able to be opened, among other subtle issue.

This is a frequently reported bug, and we try to apply this workaround
elsewhere, it's just that the .desktop file that electron-builder
generates skips the workaround we have applie e.g. in the /us/bin/pulsar
launcher script.