Remove ambiguity in all text movement method names…
…of the form `(select|delete|cut|move)To(Beginning|End)OfLine`.

Also add most permutations of the form `(select|delete|cut|move)To(Beginning|End)Of(Screen|Buffer)Line`.

The ambiguous names are the ones that do not specify whether they operate on buffer coordinates or screen coordinates, and their behavior is inconsistent as a result. The behavior of each existing ambiguous method is preserved; the old method names are now aliases for the new ones.

This is a work in progress; the same work should be done in aliasing the respective command names that map to the old methods, plus creating new command names for the new methods.