Use first significant content on the line (not column 0)…
…when determining suggested indent.

There's a lot of nuance here about which layer gets asked the question of what the current line's indentation level should be. To make sure we're asking the right layer, we want to consider the deepest layer that is active at a given point.

But since injections rarely cover the whitespace that precedes them, we generally want to look at the beginnings and ends of the _content_ on various lines. So if the first ten characters of a line are whitespace, we should ask what the controlling layer is at column 10, not at column 0.