electron-builder: Fix race condition when preparing to copy binaries
There was a race condition where two separate pieces of async code
would A) create the 'binaries' folder, and B) copy binaries to that
folder. The problem was the lack of a guarantee they would happen
in that order! If the attempt to copy files happened before
the directory got created, the whole electron-builder run would
error and/or stall out from unsuccessful auto-retries failing.

To fix the situation:

Ensure the 'binaries' dir exists *before* copying binaries to it.

Do so by making the directory *first* with a *synchronous* function.

(I might usually try to eliminate race conditions while still using
asynchronous code, but for simple scripts, I prefer to use
synchronous code, for simplicity's sake and for reliability.)

In other words: Can't copy files to a destination dir that
doesn't exist! This makes sure the directory exists *first*, then copy.

Also: Log any error we hit while making the dir, but try to continue.
(For anyone who has successfully run electorn-builder once in the repo,
they already have a 'binaries' folder, so we can ignore any error when
trying to "make the dir" that already exists. If it doesn't exist, the
copy operation will print another relevant error message. So we don't
need to error early here. But logging it should help if this issue
needs any troubleshooting on anyone's machines (including CI) any time
in the future.)