CI: Retry on timeout when building macOS bins
Also, adjust the timeout duration to be shorter,
and adjust the number of retries to be greater.

I figure successful runs usually succeed quickly (~27 minutes or less?)
So, get to the retrying faster! But also, let's actually retry,
a bunch more times if we have to!

(cherry picked from commit a1af458c9edb07975230219b2a46557857651503)

  Conflicts:
 	.github/workflows/build.yml