CI: Reduce macOS bins timeout to 31 minutes
37 minutes is a bit long. 31 minutes should allow most runs that are
*going* to finish to finish, without wasting more time than we need to
waiting for runs that are hung and are never going to finish anyway.

Considered 27, since *most* runs finish at 16 minutes, and all but one
of my ~11 test runs were under 27 minutes, but *one* of my test runs was
just under 31 minutes, so this should do the trick.

Having to wait for an entire additional run by being over-eager to retry
(too impatient) feels wasteful and not the best trade-off.

Anyhow, this is closer to the right number, even if picking a better
"exact ideal timeout" may be possible, this is close enough.